testing fix profile, course, section, subSection update and delete & mail change mail address & mailsender remove log

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require('../models/SubSection');
 
 //create section
 exports.createSection = async (req, res) => {
@@ -83,25 +84,47 @@ exports.updateSection = async (req, res) => {
 
 // delete section
 exports.deleteSection = async (req, res) => {
-    try {
-      //data input
-      const {sectionId } = req.params
-  
-      //delete data
-      await Section.findByIdAndDelete(sectionId);
-      //Too(testing time): do we need to delete the entry from the course schema?
-  
-      //return response
-      return res.status(200).json({
-        success: true,
-        message: "Section delete Successfully",
-      });
-    } catch (error) {
-      console.log(error);
-      return res.status(500).json({
+  try {
+    const { sectionId, courseId } = req.body;
+    await Course.findByIdAndUpdate(courseId, {
+      $pull: {
+        courseContent: sectionId,
+      },
+    });
+    const section = await Section.findById(sectionId);
+    console.log(sectionId, courseId);
+    if (!section) {
+      return res.status(404).json({
         success: false,
-        message: "Unable to Update sections, please try again",
-        error: error.message,
+        message: "Section not Found",
       });
     }
-  };
+
+    //delete sub section
+    await SubSection.deleteMany({ _id: { $in: section.subSection } });
+
+    await Section.findByIdAndDelete(sectionId);
+
+    //find the updated course and return
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec();
+
+    res.status(200).json({
+      success: true,
+      message: "Section deleted",
+      data: course,
+    });
+  } catch (error) {
+    console.error("Error deleting section:", error);
+    res.status(500).json({
+      success: false,
+      message: "Internal server error",
+    });
+  }
+};
